move to util function
